🔒 chore(userController.js): add token generation and authentication logic 🔒 chore(userModel.js): comment out password field selection 📦 chore(package.json): add jsonwebtoken dependency
In the userController.js file, token generation and authentication logic have been added. The registerUser function now generates a token using the gernerateToken function and includes it in the response. The loginUser function now checks if the provided email and password match and generates a token if they do. The gernerateToken function is a new function that generates a token using the jsonwebtoken library.

In the userModel.js file, the select option for the password field has been commented out. This means that the password field will be included in query results by default.

In the package.json file, the jsonwebtoken dependency has been added to the dependencies section.